+++ internal/assets-service/services/cosmetics/service.go
@@ -21,11 +21,12 @@ type CosmeticsService interface {
 	GetCosmeticsListByWorld(worldId uuid.UUID, offset int, limit int) ([]*models.Cosmetic, int64, error)
 
 	// UploadCosmeticData handles the upload of cosmetic file.
-	UploadCosmeticData(category uuid.UUID, worldId uuid.UUID, price float64, cosmeticData multipart.File, ext string, userId uuid.UUID) (*models.Cosmetic, error)
+	UploadCosmeticData(category uuid.UUID, worldId uuid.UUID, price int64, cosmeticData multipart.File, ext string, userId uuid.UUID) (*models.Cosmetic, error)
 
 	// UploadCosmeticByID links an existing cosmetic sprite to another category,
 	// optionally replacing the sprite file without changing its ID.
-	UploadCosmeticByID(categoryId uuid.UUID, worldId uuid.UUID, price float64, spriteId uuid.UUID, userId uuid.UUID, cosmeticFile multipart.File, ext string) (*models.Cosmetic, error)
+	UploadCosmeticByID(categoryId uuid.UUID, worldId uuid.UUID, price int64, spriteId uuid.UUID, userId uuid.UUID, cosmeticFile multipart.File, ext string) (*models.Cosmetic, error)
+
 	// DeleteCosmetic handles the deletion of a cosmetic by its ID.
 	DeleteCosmetic(cosmeticId uuid.UUID) error
