+++ internal/payment-service/repositories/gem-balances/repository.go
@@ -9,7 +9,7 @@ type GemBalancesRepository interface {
 	CreateGemBalance(userId uuid.UUID) error
 	GetAllGemBalances() ([]*models.GemBalance, error)
 	GetGemBalanceByUserId(userId uuid.UUID) (*models.GemBalance, error)
-	AddToGemBalance(userId uuid.UUID, gems int) error
-	ApplyStripeCheckoutCreditIfUnprocessed(userId uuid.UUID, gems int, eventID string, sessionID string) (bool, error)
-	UpdateGemBalance(userId uuid.UUID, gems int) error
+	AddToGemBalance(userId uuid.UUID, gems int64) error
+	ApplyStripeCheckoutCreditIfUnprocessed(userId uuid.UUID, gems int64, eventID string, sessionID string) (bool, error)
+	UpdateGemBalance(userId uuid.UUID, gems int64) error
 }
